move xml parser to utils folder
